added normalisation to vector (does not work correctly), added static cast to velocity to remove warnings (do not want any runtime checks), refactored collidable tiles to use the Collidable class.

+++ DauntingDungeon/InputManager.cpp
@@ -10,8 +10,8 @@ void InputManager::HandleInput(bool &isRunning, Vector2 &playerVel)
 	SDL_PollEvent(&event);
 
 	const Uint8* keyState = SDL_GetKeyboardState(NULL);
-	playerVel.x = keyState[SDL_SCANCODE_D] - keyState[SDL_SCANCODE_A];
-	playerVel.y = keyState[SDL_SCANCODE_S] - keyState[SDL_SCANCODE_W];
+	playerVel.x = static_cast<float>(keyState[SDL_SCANCODE_D] - keyState[SDL_SCANCODE_A]);
+	playerVel.y = static_cast<float>(keyState[SDL_SCANCODE_S] - keyState[SDL_SCANCODE_W]);
 
 	if (event.type == SDL_KEYDOWN) {
 		switch (event.key.keysym.sym) {
